+++ gamemodes/terrortown/gamemode/server/sv_weaponry.lua
@@ -13,6 +13,7 @@ local IsValid = IsValid
 local table = table
 local timer = timer
 local hook = hook
+local playerIterator = player.Iterator
 
 local IsEquipment = WEPS.IsEquipment
 
@@ -470,7 +471,7 @@ end
 -- @realm server
 -- @ref https://wiki.facepunch.com/gmod/GM:PlayerLoadout
 function GM:UpdatePlayerLoadouts()
-    local plys = player.GetAll()
+    local plys = select(2, playerIterator())
 
     for i = 1, #plys do
         ---