+++ lua/ttt2/libraries/credits.lua
@@ -66,7 +66,7 @@ function credits.HandleKillCreditsAward(victim, attacker)
 
     -- This is a special case scenario where for every kill it is checked for every player,
     -- how many players are dead from different teams than their own team.
-    local plys = player.GetAll()
+    local plys = select(2, player.Iterator())
     local plysAmount = #plys
     local plysByTeams = {}