+++ lua/ttt2/libraries/marker_vision.lua
@@ -20,6 +20,7 @@ VISIBLE_FOR_BITS = 3
 markerVision = {}
 
 markerVision.registry = {}
+markerVision.focussedMarkers = {}
 
 ---
 -- Creates a new marker vision object for the entity.
@@ -223,9 +224,31 @@ if CLIENT then
         { font = "Tahoma", size = 20, weight = 600, extended = true }
     )
     surface.CreateAdvancedFont(
+        "RadarVision_Subtitle",
+        { font = "Tahoma", size = 17, weight = 300, extended = true }
+    )
+    surface.CreateAdvancedFont(
         "RadarVision_Text",
         { font = "Tahoma", size = 14, weight = 300, extended = true }
     )
+
+    ---
+    -- This gets the currently focussed closest entity.
+    -- @realm client
+    function markerVision.GetFocusedEntity()
+        local closestEntTbl = markerVision.focussedMarkers[1] or {}
+
+        for i = 2, #markerVision.focussedMarkers do
+            local entTbl = markerVision.focussedMarkers[i]
+            if entTbl.screenDistanceSquared < closestEntTbl.screenDistanceSquared then
+                closestEntTbl = entTbl
+            end
+        end
+
+        markerVision.focussedMarkers = { closestEntTbl }
+
+        return closestEntTbl.entity
+    end
 
     ---
     -- The draw function of the radar vision module.
@@ -234,7 +257,7 @@ if CLIENT then
     function markerVision.Draw()
         local scale = appearance.GetGlobalScale()
 
-        local padding = 3 * scale
+        local padding = 4 * scale
         local paddingScreen = 10 * scale
 
         local sizeIcon = 28 * scale
@@ -246,6 +269,11 @@ if CLIENT then
         local sizeTitleIcon = 14 * scale
         local offsetTitleIcon = 0.5 * sizeTitleIcon
 
+        local sizeSubtitleIcon = 12 * scale
+        local offsetSubtitleIcon = 0.5 * sizeSubtitleIcon
+
+        local offsetSubtitle = 8 * scale
+
         local heightLineDescription = 14 * scale
 
         local client = LocalPlayer()
@@ -253,6 +281,8 @@ if CLIENT then
         local heightScreen = ScrH()
         local xScreenCenter = 0.5 * widthScreen
         local yScreenCenter = 0.5 * heightScreen
+
+        markerVision.focussedMarkers = {}
 
         for i = 1, #markerVision.registry do
             local mvObject = markerVision.registry[i]
@@ -265,11 +295,9 @@ if CLIENT then
             local posEnt = ent:GetPos() + ent:OBBCenter()
             local screenPos = posEnt:ToScreen()
             local isOffScreen = util.IsOffScreen(screenPos)
-            local isOnScreenCenter = not isOffScreen
-                and screenPos.x > xScreenCenter - offsetIcon
-                and screenPos.x < xScreenCenter + offsetIcon
-                and screenPos.y > yScreenCenter - offsetIcon
-                and screenPos.y < yScreenCenter + offsetIcon
+            local screenDistSquared = (screenPos.x - xScreenCenter) ^ 2
+                + (screenPos.y - yScreenCenter) ^ 2
+            local isOnScreenCenter = not isOffScreen and screenDistSquared < offsetIcon ^ 2
             local distanceEntity = posEnt:Distance(client:EyePos())
 
             -- call internal targetID functions first so the data can be modified by addons
@@ -292,6 +320,13 @@ if CLIENT then
 
             if not params.drawInfo then
                 continue
+            end
+
+            if isOnScreenCenter then
+                markerVision.focussedMarkers[#markerVision.focussedMarkers + 1] = {
+                    entity = ent,
+                    screenDistanceSquared = screenDistSquared,
+                }
             end
 
             local amountIcons = #params.displayInfo.icon
@@ -352,15 +387,19 @@ if CLIENT then
                 continue
             end
 
-            screenPos.x = math.Clamp(
-                screenPos.x,
-                offsetIcon + paddingScreen,
-                widthScreen - offsetIcon - paddingScreen
+            screenPos.x = math.Round(
+                math.Clamp(
+                    screenPos.x,
+                    offsetIcon + paddingScreen,
+                    widthScreen - offsetIcon - paddingScreen
+                )
             )
-            screenPos.y = math.Clamp(
-                screenPos.y,
-                offsetIcon + paddingScreen,
-                heightScreen - offsetIcon - paddingScreen
+            screenPos.y = math.Round(
+                math.Clamp(
+                    screenPos.y,
+                    offsetIcon + paddingScreen,
+                    heightScreen - offsetIcon - paddingScreen
+                )
             )
 
             -- draw Icons
@@ -388,11 +427,14 @@ if CLIENT then
                 end
             end
 
+            -- check if a subtitle is set because that shifts multiple things around
+            local hasSubtitle = mvData:HasSubtitle()
+
             -- draw title
             local stringTitle = params.displayInfo.title.text
 
             local xStringTitle = screenPos.x + offsetIcon + padding
-            local yStringTitle = screenPos.y
+            local yStringTitle = hasSubtitle and (screenPos.y - offsetSubtitle) or screenPos.y
 
             for j = 1, #params.displayInfo.title.icons do
                 drawsc.FilteredShadowedTexture(
@@ -405,7 +447,7 @@ if CLIENT then
                     params.displayInfo.title.color
                 )
 
-                xStringTitle = xStringTitle + 18 * scale
+                xStringTitle = xStringTitle + sizeTitleIcon + 4 * scale
             end
 
             draw.AdvancedText(
@@ -420,6 +462,38 @@ if CLIENT then
                 scale
             )
 
+            -- draw subtitle
+            if hasSubtitle then
+                local stringSubtitle = params.displayInfo.subtitle.text
+
+                local xStringSubtitle = xStringTitle
+                local yStringSubtitle = screenPos.y + offsetSubtitle
+
+                for j = 1, #params.displayInfo.subtitle.icons do
+                    drawsc.FilteredShadowedTexture(
+                        xStringSubtitle,
+                        yStringSubtitle - offsetSubtitleIcon,
+                        sizeSubtitleIcon,
+                        sizeSubtitleIcon,
+                        params.displayInfo.subtitle.icons[j],
+                        params.displayInfo.subtitle.color.a,
+                        params.displayInfo.subtitle.color
+                    )
+
+                    xStringSubtitle = xStringSubtitle + sizeSubtitleIcon + 4 * scale
+                end
+
+                drawsc.AdvancedShadowedText(
+                    stringSubtitle,
+                    "RadarVision_Subtitle",
+                    xStringSubtitle,
+                    yStringSubtitle,
+                    params.displayInfo.subtitle.color,
+                    TEXT_ALIGN_LEFT,
+                    TEXT_ALIGN_CENTER
+                )
+            end
+
             -- draw description
             local linesDescription = params.displayInfo.desc
             local amountLinesDescription = #linesDescription
@@ -432,7 +506,9 @@ if CLIENT then
                 local color = linesDescription[j].color
 
                 local xStringDescriptionShifted = xStringDescription
-                local yStringDescription = yStringTitle + j * heightLineDescription
+                local yStringDescription = yStringTitle
+                    + j * heightLineDescription
+                    + (hasSubtitle and 3 * offsetSubtitle or 0)
 
                 for k = 1, #icons do
                     draw.FilteredShadowedTexture(